feat: add support for using teslamate geofences (#12)
Allows using Geofences defined in TeslaMate rather than manually
specifying a lat/long center point with radii. Some observations
regarding this method:

* Geo does not update as frequently as lat/long in MQTT, so there is
quite a delay in actually triggering the garage action
* Presumably this is because the Geofence isn't calculated in realtime
with every lat/long update like Tesla-YouQ does, but this is an
assumption
* TeslaMate doesn't seem to have any logic for distinction between
overlapping geofences. For example, with 2 geofences that have the same
center point but different radii, TeslaMate exhibits unpredictable
behavior with which geofence gets ascribed to the location
* This limitation means this method should really only be used for a
single geofence, meaning open and close events cannot be defined
separately
* This PR will still support having separate geo's but is at the mercy
of the data it receives from TeslaMate; if this is ever fixed, then
Tesla-YouQ will already support it, but in the meantime, users should
stick to a single geofence
* This method requires defining a `DEFAULT_GEOFENCE` environment
variable for TeslaMate (such as `not_home`), as some string must be
published to the MQTT broker to trigger an event.
* See [this PR](https://github.com/adriankumpf/teslamate/pull/2564) in
TeslaMate for more info